refactor: refactor downloader to split hugging face and modelscope download logics
