+++ apps/Android/MnnLlmChat/app/src/main/java/com/alibaba/mls/api/download/DownloadPersistentData.kt
@@ -15,8 +15,7 @@ object DownloadPersistentData {
 
     const val SIZE_SAVED_KEY: String = "size_saved"
 
-
-    fun saveMetaData(context: Context, modelId: String?, metaDataList: List<HfFileMetadata?>?) {
+    fun saveMetaData(context: Context, modelId: String, metaDataList: List<HfFileMetadata?>?) {
         var modelId = modelId
         modelId = getLastFileName(modelId)
         val sharedPreferences =
@@ -28,8 +27,7 @@ object DownloadPersistentData {
         editor.apply()
     }
 
-    @JvmStatic
-    fun getMetaData(context: Context, modelId: String?): List<HfFileMetadata>? {
+    fun getMetaData(context: Context, modelId: String): List<HfFileMetadata>? {
         var modelId = modelId
         modelId = getLastFileName(modelId)
         val sharedPreferences =
@@ -43,8 +41,7 @@ object DownloadPersistentData {
         return null
     }
 
-    @JvmStatic
-    fun saveDownloadSizeTotal(context: Context, modelId: String?, total: Long) {
+    fun saveDownloadSizeTotal(context: Context, modelId: String, total: Long) {
         var modelId = modelId
         modelId = getLastFileName(modelId)
         val sharedPreferences =
@@ -55,7 +52,7 @@ object DownloadPersistentData {
     }
 
     @JvmStatic
-    fun getDownloadSizeTotal(context: Context, modelId: String?): Long {
+    fun getDownloadSizeTotal(context: Context, modelId: String): Long {
         var modelId = modelId
         modelId = getLastFileName(modelId)
         val sharedPreferences =
@@ -64,7 +61,7 @@ object DownloadPersistentData {
     }
 
     @JvmStatic
-    fun saveDownloadSizeSaved(context: Context, modelId: String?, saved: Long) {
+    fun saveDownloadSizeSaved(context: Context, modelId: String, saved: Long) {
         var modelId = modelId
         modelId = getLastFileName(modelId)
         val sharedPreferences =
@@ -74,8 +71,7 @@ object DownloadPersistentData {
         editor.apply()
     }
 
-    @JvmStatic
-    fun getDownloadSizeSaved(context: Context, modelId: String?): Long {
+    fun getDownloadSizeSaved(context: Context, modelId: String): Long {
         var modelId = modelId
         modelId = getLastFileName(modelId)
         val sharedPreferences =
@@ -83,8 +79,7 @@ object DownloadPersistentData {
         return sharedPreferences.getLong(SIZE_SAVED_KEY, 0)
     }
 
-    @JvmStatic
-    fun removeProgress(context: Context, modelId: String?) {
+    fun removeProgress(context: Context, modelId: String) {
         var modelId = modelId
         modelId = getLastFileName(modelId)
         val sharedPreferences =
