+++ apps/Android/MnnLlmChat/app/src/main/java/com/alibaba/mls/api/download/ModelDownloadManager.kt
@@ -15,70 +15,35 @@ import androidx.core.content.ContextCompat
 import com.alibaba.mls.api.ApplicationProvider
 import com.alibaba.mls.api.HfApiClient
 import com.alibaba.mls.api.HfApiClient.Companion.bestClient
-import com.alibaba.mls.api.HfApiClient.RepoInfoCallback
-import com.alibaba.mls.api.HfApiException
-import com.alibaba.mls.api.HfFileMetadata
-import com.alibaba.mls.api.HfRepoInfo
-import com.alibaba.mls.api.download.DownloadExecutor.Companion.executor
-import com.alibaba.mls.api.download.DownloadFileUtils.createSymlink
-import com.alibaba.mls.api.download.DownloadFileUtils.deleteDirectoryRecursively
 import com.alibaba.mls.api.download.DownloadFileUtils.getLastFileName
-import com.alibaba.mls.api.download.DownloadFileUtils.getPointerPathParent
-import com.alibaba.mls.api.download.DownloadFileUtils.repoFolderName
 import com.alibaba.mls.api.download.DownloadPersistentData.getDownloadSizeSaved
 import com.alibaba.mls.api.download.DownloadPersistentData.getDownloadSizeTotal
-import com.alibaba.mls.api.download.DownloadPersistentData.getMetaData
 import com.alibaba.mls.api.download.DownloadPersistentData.removeProgress
 import com.alibaba.mls.api.download.DownloadPersistentData.saveDownloadSizeSaved
 import com.alibaba.mls.api.download.DownloadPersistentData.saveDownloadSizeTotal
-import com.alibaba.mls.api.download.DownloadPersistentData.saveMetaData
-import com.alibaba.mls.api.download.HfFileMetadataUtils.getFileMetadata
-import com.alibaba.mls.api.download.ModelFileDownloader.FileDownloadListener
 import com.alibaba.mls.api.ms.MsApiClient
-import com.alibaba.mls.api.ms.MsRepoInfo
 import com.alibaba.mls.api.source.ModelSources
-import com.alibaba.mls.api.source.RepoConfig
 import com.alibaba.mnnllm.android.utils.FileUtils.clearMmapCache
-import okhttp3.OkHttpClient
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 import java.io.File
-import java.util.Collections
-import java.util.concurrent.TimeUnit
 import java.util.concurrent.atomic.AtomicInteger
-import kotlin.collections.ArrayList
-import kotlin.collections.HashMap
-import kotlin.collections.HashSet
-import kotlin.collections.List
-import kotlin.collections.MutableList
-import kotlin.collections.MutableSet
-import kotlin.collections.containsKey
-import kotlin.collections.dropLastWhile
-import kotlin.collections.get
-import kotlin.collections.indices
 import kotlin.collections.set
-import kotlin.collections.toTypedArray
 import kotlin.concurrent.Volatile
 
 class ModelDownloadManager private constructor(private val context: Context) {
     private var downloadListener: DownloadListener? = null
     private val cachePath = context.filesDir.absolutePath + "/.mnnmodels"
-    private val modelScopeCachePath = cachePath + "/modelscope"
-
+    val modelScopeCachePath = "$cachePath/modelscope"
+    private var downloader: ModelRepoDownloader? = null
     private var hfApiClient: HfApiClient? = null
-    private var msApiClient: MsApiClient? = null
+    var msApiClient: MsApiClient? = null
         get() {
             if (field == null) {
                 field = MsApiClient()
             }
             return field
         }
-
-    private var metaInfoClient: OkHttpClient? = null
     private val downloadInfoMap = HashMap<String, DownloadInfo>()
-    private val pausedSet: MutableSet<String?> = Collections.synchronizedSet(HashSet())
-    private val foregroundSerivceIntent =
+    private val foregroundServiceIntent =
         Intent(context.applicationContext, DownlodForegroundService::class.java)
 
     private val activeDownloadCount =
@@ -90,11 +55,54 @@ class ModelDownloadManager private constructor(private val context: Context) {
         this.downloadListener = downloadListener
     }
 
-    fun getDownloadPath(modelId: String): File {
+    private fun getDownloadPath(modelId: String): File {
         return if (ModelSources.get().remoteSourceType == ModelSources.ModelSourceType.HUGGING_FACE) {
             getHfDownloadModelPath(modelId)
         } else {
             getMsModelPath(modelId)
+        }
+    }
+
+    private fun setDownloader(sourceType: ModelSources.ModelSourceType) {
+        val downloadCallback = object : ModelRepoDownloader.ModelRepoDownloadCallback {
+            override fun onDownloadFailed(modelId: String, e: Exception) {
+                setDownloadFailed(modelId, e)
+            }
+
+            override fun onDownloadTaskAdded() {
+                onDownloadTaskAdded(activeDownloadCount.incrementAndGet())
+            }
+
+            override fun onDownloadTaskRemoved() {
+                onDownloadTaskRemoved(activeDownloadCount.decrementAndGet())
+            }
+
+            override fun onDownloadPaused(modelId: String) {
+                setDownloadPaused(modelId)
+            }
+
+            override fun onDownloadFileFinished(modelId: String, absolutePath: String) {
+                setDownloadFinished(modelId, absolutePath)
+            }
+
+            override fun onDownloadingProgress(
+                modelId: String,
+                stage: String,
+                currentFile: String?,
+                saved: Long,
+                total: Long
+            ) {
+                updateDownloadingProgress(
+                    modelId, "file", currentFile,
+                    saved, total
+                )
+            }
+        }
+        this.downloader = when (sourceType) {
+            ModelSources.ModelSourceType.HUGGING_FACE -> HuggingFaceModelDownloader(this, downloadCallback, cachePath)
+            ModelSources.ModelSourceType.MODELERS -> ModelScopeModelDownloader(this, downloadCallback, this.modelScopeCachePath)
+            ModelSources.ModelSourceType.MODEL_SCOPE -> ModelScopeModelDownloader(this, downloadCallback, this.modelScopeCachePath)
+            ModelSources.ModelSourceType.LOCAL -> null
         }
     }
 
@@ -110,11 +118,11 @@ class ModelDownloadManager private constructor(private val context: Context) {
         return null
     }
 
-    private fun getHfDownloadModelPath(modelId: String): File {
+    fun getHfDownloadModelPath(modelId: String): File {
         return File(cachePath, getLastFileName(modelId))
     }
 
-    private fun getMsModelPath(modelId: String): File {
+    fun getMsModelPath(modelId: String): File {
         val modelScopeId = ModelSources.get().config.getRepoConfig(modelId)!!.modelScopePath
         return File(this.modelScopeCachePath, getLastFileName(modelScopeId))
     }
@@ -123,10 +131,10 @@ class ModelDownloadManager private constructor(private val context: Context) {
         if (getDownloadInfo(modelId).downlodaState != DownloadInfo.DownloadSate.DOWNLOADING) {
             return
         }
-        pausedSet.add(modelId)
+        downloader?.pause(modelId)
     }
 
-    private fun getHfApiClient(): HfApiClient {
+     fun getHfApiClient(): HfApiClient {
         if (hfApiClient == null) {
             hfApiClient = bestClient
         }
@@ -141,49 +149,8 @@ class ModelDownloadManager private constructor(private val context: Context) {
             downloadListener!!.onDownloadStart(modelId)
         }
         this.updateDownloadingProgress(modelId, "Preparing", null, 0, 10)
-        if (ModelSources.get().remoteSourceType == ModelSources.ModelSourceType.HUGGING_FACE) {
-            getHfApiClient().getRepoInfo(modelId, "main", object : RepoInfoCallback {
-                override fun onSuccess(hfRepoInfo: HfRepoInfo?) {
-                    downloadHfRepo(hfRepoInfo!!)
-                }
-
-                override fun onFailure(error: String?) {
-                    setDownloadFailed(modelId, HfApiException("getRepoInfoFailed$error"))
-                }
-            })
-        } else {
-            downloadMsRepo(modelId)
-        }
-    }
-
-    private fun downloadMsRepo(modelId: String) {
-        val repoConfig = ModelSources.get().config.getRepoConfig(modelId)
-        val modelScopeId = repoConfig!!.repositoryPath()
-        val split = modelScopeId.split("/".toRegex()).dropLastWhile { it.isEmpty() }.toTypedArray()
-        if (split.size != 2) {
-            setDownloadFailed(
-                modelId,
-                HfApiException("getRepoInfoFailed modelId format error: $modelId")
-            )
-        }
-        msApiClient!!.apiService.getModelFiles(split[0], split[1])
-            .enqueue(object : Callback<MsRepoInfo?> {
-                override fun onResponse(call: Call<MsRepoInfo?>, response: Response<MsRepoInfo?>) {
-                    executor!!.submit {
-                        Log.d(
-                            TAG,
-                            "downloadMsRepoInner executor"
-                        )
-                        onDownloadTaskAdded(activeDownloadCount.incrementAndGet())
-                        downloadMsRepoInner(repoConfig, response.body()!!)
-                        onDownloadTaskRemoved(activeDownloadCount.decrementAndGet())
-                    }
-                }
-
-                override fun onFailure(call: Call<MsRepoInfo?>, t: Throwable) {
-                    setDownloadFailed(modelId, HfApiException("getRepoInfoFailed" + t.message))
-                }
-            })
+        setDownloader(ModelSources.get().remoteSourceType)
+        downloader?.download(modelId)
     }
 
     @SuppressLint("DefaultLocale")
@@ -208,73 +175,6 @@ class ModelDownloadManager private constructor(private val context: Context) {
             } else {
                 downloadInfo.speedInfo = String.format("%.2fB/s", speedBytesPerSecond)
             }
-        }
-    }
-
-    private fun downloadMsRepoInner(repoConfig: RepoConfig, msRepoInfo: MsRepoInfo) {
-        Log.d(TAG, "downloadMsRepoInner")
-        val modelId = repoConfig.modelId
-        val folderLinkFile =
-            File(this.modelScopeCachePath, getLastFileName(repoConfig.repositoryPath()))
-        if (folderLinkFile.exists()) {
-            Log.d(TAG, "downloadMsRepoInner already exists")
-            setDownloadFinished(repoConfig.modelId, folderLinkFile.absolutePath)
-            return
-        }
-        val modelDownloader = ModelFileDownloader()
-        val hasError = false
-        val errorInfo = StringBuilder()
-        val repoFolderName = repoFolderName(repoConfig.repositoryPath(), "model")
-        val storageFolder = File(this.modelScopeCachePath, repoFolderName)
-        val parentPointerPath = getPointerPathParent(storageFolder, "_no_sha_")
-        val downloadTaskList: List<FileDownloadTask>
-        val totalAndDownloadSize = LongArray(2)
-        Log.d(TAG, "downloadMsRepoInner collectMsTaskList")
-        downloadTaskList = collectMsTaskList(
-            repoConfig,
-            storageFolder,
-            parentPointerPath,
-            msRepoInfo,
-            totalAndDownloadSize
-        )
-        Log.d(TAG, "downloadMsRepoInner downloadTaskList： " + downloadTaskList.size)
-        val fileDownloadListener =
-            object :FileDownloadListener {
-                override fun onDownloadDelta(
-                    fileName: String?,
-                    downloadedBytes: Long,
-                    totalBytes: Long,
-                    delta: Long
-                ): Boolean {
-                    totalAndDownloadSize[1] += delta
-                    updateDownloadingProgress(
-                        modelId, "file", fileName,
-                        totalAndDownloadSize[1], totalAndDownloadSize[0]
-                    )
-                    return pausedSet.contains(modelId)
-                }
-            }
-        try {
-            for (fileDownloadTask in downloadTaskList) {
-                modelDownloader.downloadFile(fileDownloadTask, fileDownloadListener)
-            }
-        } catch (e: DownloadPausedException) {
-            pausedSet.remove(modelId)
-            setDownloadPaused(modelId)
-            return
-        } catch (e: Exception) {
-            setDownloadFailed(modelId, e)
-            return
-        }
-        if (!hasError) {
-            val folderLinkPath = folderLinkFile.absolutePath
-            createSymlink(parentPointerPath.toString(), folderLinkPath)
-            setDownloadFinished(modelId, folderLinkPath)
-        } else {
-            Log.e(
-                TAG,
-                "Errors occurred during download: $errorInfo"
-            )
         }
     }
 
@@ -331,7 +231,7 @@ class ModelDownloadManager private constructor(private val context: Context) {
                         REQUEST_CODE_POST_NOTIFICATIONS
                     )
                 } else {
-                    ApplicationProvider.get().startForegroundService(foregroundSerivceIntent)
+                    ApplicationProvider.get().startForegroundService(foregroundServiceIntent)
                     foregroundServiceStarted = true
                 }
             }
@@ -340,12 +240,12 @@ class ModelDownloadManager private constructor(private val context: Context) {
 
     private fun onDownloadTaskRemoved(count: Int) {
         if (count == 0) {
-            ApplicationProvider.get().stopService(foregroundSerivceIntent)
+            ApplicationProvider.get().stopService(foregroundServiceIntent)
             foregroundServiceStarted = false
         }
     }
 
-    private fun setDownloadFailed(modelId: String, e: Exception) {
+    fun setDownloadFailed(modelId: String, e: Exception) {
         Log.e(TAG, "onDownloadFailed: $modelId", e)
         val downloadInfo = getDownloadInfo(modelId)
         downloadInfo.downlodaState = DownloadInfo.DownloadSate.FAILED
@@ -356,7 +256,7 @@ class ModelDownloadManager private constructor(private val context: Context) {
     }
 
     private fun updateDownloadingProgress(
-        modelId: String?,
+        modelId: String,
         stage: String,
         currentFile: String?,
         saved: Long,
@@ -364,7 +264,7 @@ class ModelDownloadManager private constructor(private val context: Context) {
     ) {
         if (!downloadInfoMap.containsKey(modelId)) {
             val downloadInfo = DownloadInfo()
-            downloadInfoMap[modelId!!] = downloadInfo
+            downloadInfoMap[modelId] = downloadInfo
         }
         val downloadInfo = checkNotNull(downloadInfoMap[modelId])
         downloadInfo.progress = saved.toDouble() / total
@@ -380,213 +280,10 @@ class ModelDownloadManager private constructor(private val context: Context) {
         }
     }
 
-    private val metaInfoHttpClient: OkHttpClient
-        get() {
-            if (metaInfoClient == null) {
-                metaInfoClient = OkHttpClient.Builder()
-                    .connectTimeout(30, TimeUnit.SECONDS)
-                    .followRedirects(false)
-                    .followSslRedirects(false)
-                    .build()
-            }
-            return metaInfoClient!!
-        }
-
-    @Throws(HfApiException::class)
-    private fun requestMedataDataList(hfRepoInfo: HfRepoInfo): List<HfFileMetadata?> {
-        val list: MutableList<HfFileMetadata?> = ArrayList()
-        for (subFile in hfRepoInfo.getSiblings()) {
-            val url =
-                "https://" + hfApiClient!!.host + "/" + hfRepoInfo.modelId + "/resolve/main/" + subFile.rfilename
-            val metaData = getFileMetadata(metaInfoHttpClient, url)
-            list.add(metaData)
-        }
-        return list
-    }
-
-    private fun collectMsTaskList(
-        repoConfig: RepoConfig,
-        storageFolder: File,
-        parentPointerPath: File,
-        msRepoInfo: MsRepoInfo,
-        totalAndDownloadSize: LongArray
-    ): List<FileDownloadTask> {
-        val fileDownloadTasks: MutableList<FileDownloadTask> = ArrayList()
-        for (i in msRepoInfo.Data.Files.indices) {
-            val subFile = msRepoInfo.Data.Files[i]
-            val fileDownloadTask = FileDownloadTask()
-            fileDownloadTask.relativePath = subFile.Path
-            fileDownloadTask.hfFileMetadata = HfFileMetadata()
-            if (ModelSources.get().remoteSourceType == ModelSources.ModelSourceType.MODELERS) {
-                fileDownloadTask.hfFileMetadata!!.location = String.format(
-                    "https://modelers.cn/coderepo/web/v1/file/%s/main/media/%s",
-                    repoConfig.repositoryPath(),
-                    subFile.Path
-                )
-            } else {
-                fileDownloadTask.hfFileMetadata!!.location = String.format(
-                    "https://modelscope.cn/api/v1/models/%s/repo?FilePath=%s",
-                    repoConfig.repositoryPath(),
-                    subFile.Path
-                )
-            }
-            fileDownloadTask.hfFileMetadata!!.size = subFile.Size
-            fileDownloadTask.hfFileMetadata!!.etag = subFile.Sha256
-            fileDownloadTask.blobPath = File(storageFolder, "blobs/" + subFile.Sha256)
-            fileDownloadTask.blobPathIncomplete =
-                File(storageFolder, "blobs/" + subFile.Sha256 + ".incomplete")
-            fileDownloadTask.pointerPath = File(parentPointerPath, subFile.Path)
-            fileDownloadTask.downloadedSize =
-                if (fileDownloadTask.blobPath!!.exists()) fileDownloadTask.blobPath!!.length() else (if (fileDownloadTask.blobPathIncomplete!!.exists()) fileDownloadTask.blobPathIncomplete!!.length() else 0)
-            totalAndDownloadSize[0] += subFile.Size
-            totalAndDownloadSize[1] += fileDownloadTask.downloadedSize
-            fileDownloadTasks.add(fileDownloadTask)
-        }
-        return fileDownloadTasks
-    }
-
-    @Throws(HfApiException::class)
-    private fun collectTaskList(
-        storageFolder: File,
-        parentPointerPath: File,
-        hfRepoInfo: HfRepoInfo,
-        totalAndDownloadSize: LongArray
-    ): List<FileDownloadTask> {
-        var metaData: HfFileMetadata
-        var metaDataList: List<HfFileMetadata?>? =
-            getMetaData(ApplicationProvider.get(), hfRepoInfo.modelId)
-        Log.d(
-            TAG, "collectTaskList savedMetaDataList: " + (metaDataList?.size
-                ?: "null")
-        )
-        val fileDownloadTasks: MutableList<FileDownloadTask> = ArrayList()
-        metaDataList = requestMedataDataList(hfRepoInfo)
-        saveMetaData(ApplicationProvider.get(), hfRepoInfo.modelId, metaDataList)
-        for (i in hfRepoInfo.getSiblings().indices) {
-            val subFile = hfRepoInfo.getSiblings()[i]
-            metaData = metaDataList[i]!!
-            val fileDownloadTask = FileDownloadTask()
-            fileDownloadTask.relativePath = subFile.rfilename
-            fileDownloadTask.hfFileMetadata = metaData
-            fileDownloadTask.blobPath = File(storageFolder, "blobs/" + metaData.etag)
-            fileDownloadTask.blobPathIncomplete =
-                File(storageFolder, "blobs/" + metaData.etag + ".incomplete")
-            fileDownloadTask.pointerPath = File(parentPointerPath, subFile.rfilename)
-            fileDownloadTask.downloadedSize =
-                if (fileDownloadTask.blobPath!!.exists()) fileDownloadTask.blobPath!!.length() else (if (fileDownloadTask.blobPathIncomplete!!.exists()) fileDownloadTask.blobPathIncomplete!!.length() else 0)
-            totalAndDownloadSize[0] += metaData.size
-            totalAndDownloadSize[1] += fileDownloadTask.downloadedSize
-            fileDownloadTasks.add(fileDownloadTask)
-        }
-        return fileDownloadTasks
-    }
-
-    fun downloadHfRepo(hfRepoInfo: HfRepoInfo) {
-        Log.d(TAG, "DownloadStart " + hfRepoInfo.modelId + " host: " + getHfApiClient().host)
-        executor!!.submit {
-            onDownloadTaskAdded(activeDownloadCount.incrementAndGet())
-            downloadHfRepoInner(hfRepoInfo)
-            onDownloadTaskRemoved(activeDownloadCount.decrementAndGet())
-        }
-    }
-
-    private fun downloadHfRepoInner(hfRepoInfo: HfRepoInfo) {
-        val folderLinkFile = File(cachePath, getLastFileName(hfRepoInfo.modelId))
-        if (folderLinkFile.exists()) {
-            setDownloadFinished(hfRepoInfo.modelId!!, folderLinkFile.absolutePath)
-            return
-        }
-        val modelDownloader = ModelFileDownloader()
-        Log.d(TAG, "Repo SHA: " + hfRepoInfo.sha)
-
-        val hasError = false
-        val errorInfo = StringBuilder()
-
-        val repoFolderName = repoFolderName(hfRepoInfo.modelId, "model")
-        val storageFolder = File(cachePath, repoFolderName)
-        val parentPointerPath = getPointerPathParent(
-            storageFolder,
-            hfRepoInfo.sha!!
-        )
-        val downloadTaskList: List<FileDownloadTask>
-        val totalAndDownloadSize = LongArray(2)
-        try {
-            downloadTaskList =
-                collectTaskList(storageFolder, parentPointerPath, hfRepoInfo, totalAndDownloadSize)
-        } catch (e: HfApiException) {
-            setDownloadFailed(hfRepoInfo.modelId!!, e)
-            return
-        }
-        val fileDownloadListener =
-            object :FileDownloadListener {
-                override fun onDownloadDelta(
-                    fileName: String?,
-                    downloadedBytes: Long,
-                    totalBytes: Long,
-                    delta: Long
-                ): Boolean {
-                    totalAndDownloadSize[1] += delta
-                    updateDownloadingProgress(
-                        hfRepoInfo.modelId, "file", fileName,
-                        totalAndDownloadSize[1], totalAndDownloadSize[0]
-                    )
-                    return pausedSet.contains(hfRepoInfo.modelId)
-                }
-            }
-        try {
-            for (fileDownloadTask in downloadTaskList) {
-                modelDownloader.downloadFile(fileDownloadTask, fileDownloadListener)
-            }
-        } catch (e: DownloadPausedException) {
-            pausedSet.remove(hfRepoInfo.modelId)
-            setDownloadPaused(hfRepoInfo.modelId!!)
-            return
-        } catch (e: Exception) {
-            setDownloadFailed(hfRepoInfo.modelId!!, e)
-            return
-        }
-        if (!hasError) {
-            val folderLinkPath = folderLinkFile.absolutePath
-            createSymlink(parentPointerPath.toString(), folderLinkPath)
-            setDownloadFinished(hfRepoInfo.modelId!!, folderLinkPath)
-        } else {
-            Log.e(
-                TAG,
-                "Errors occurred during download: $errorInfo"
-            )
-        }
-    }
-
-    fun removeDownload(modelId: String) {
-        val repoFolderName = repoFolderName(modelId, "model")
-        val hfStorageFolder = File(cachePath, repoFolderName)
-        Log.d(TAG, "removeStorageFolder: " + hfStorageFolder.absolutePath)
-        if (hfStorageFolder.exists()) {
-            val result = deleteDirectoryRecursively(hfStorageFolder)
-            if (!result) {
-                Log.e(TAG, "remove storageFolder" + hfStorageFolder.absolutePath + " faield")
-            }
-        }
-        val msModelId = ModelSources.get().config.getRepoConfig(modelId)!!.modelScopePath
-        val msRepoFolderName = repoFolderName(msModelId, "model")
-        val msStorageFolder = File(this.modelScopeCachePath, msRepoFolderName)
-        Log.d(TAG, "removeStorageFolder: " + msStorageFolder.absolutePath)
-        if (msStorageFolder.exists()) {
-            val result = deleteDirectoryRecursively(msStorageFolder)
-            if (!result) {
-                Log.e(TAG, "remove storageFolder" + msStorageFolder.absolutePath + " faield")
-            }
-        }
+    fun deleteRepo(modelId: String) {
+        downloader?.deleteRepo(modelId)
         removeProgress(ApplicationProvider.get(), modelId)
-        val hfLinkFolder = this.getHfDownloadModelPath(modelId)
-        Log.d(TAG, "removeHfLinkFolder: " + hfLinkFolder.absolutePath)
-        hfLinkFolder.delete()
-
-        val msLinkFolder = this.getMsModelPath(modelId)
-        Log.d(TAG, "removeMsLinkFolder: " + msLinkFolder.absolutePath)
-        msLinkFolder.delete()
         clearMmapCache(modelId)
-
         if (downloadListener != null) {
             val downloadInfo = getDownloadInfo(modelId)
             downloadInfo.downlodaState = DownloadInfo.DownloadSate.NOT_START
@@ -627,7 +324,7 @@ class ModelDownloadManager private constructor(private val context: Context) {
 
     fun startForegroundService() {
         if (!foregroundServiceStarted && activeDownloadCount.get() > 0) {
-            ApplicationProvider.get().startForegroundService(foregroundSerivceIntent)
+            ApplicationProvider.get().startForegroundService(foregroundServiceIntent)
         }
     }
